pyproject.toml
[tool.poetry]
name = "sql-table"
version = "0.1.0"
description = ""
authors = ["sayalee damle <sayalee.damle@onepointltd.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.28.1"
langchain = "^0.0.305"
python-dotenv = "^1.0.0"
faiss-cpu = "^1.7.4"
tiktoken = "^0.5.1"
chainlit = "^0.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
